+++ Source/ansic/runnerline.h
@@ -1,5 +1,5 @@
 // runnerline.h
-// Revision 3-aug-2003
+// Revision 22-aug-2003
 
 #ifndef RUNNERLINE_H_
 
@@ -287,6 +287,7 @@ private:
 	bool do_origin ();
 	bool do_deg_rad ();
 	bool do_inverse ();
+	bool do_if_debug ();
 };
 
 #endif